requirements.txt
# --- Phase 1: Core APIs & UI ---
fastapi
uvicorn
pydantic
streamlit
requests
python-dotenv

# --- Phase 1: AI & Inference ---
langchain
langchain-google-genai
sentence-transformers
textblob

# --- Phases 2 & 3: Memory & RAG (Future-proofing) ---
chromadb
faiss-cpu
rank-bm25
spacy